+++ helper_test.go
@@ -22,6 +22,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"reflect"
 	"runtime"
 	"strings"
 	"testing"
@@ -2849,6 +2850,53 @@ func enterpriseEnabled() bool {
 // Checks to see if ENABLE_BETA is set to 1, thereby enabling tests for beta features.
 func betaFeaturesEnabled() bool {
 	return os.Getenv("ENABLE_BETA") == "1"
+}
+
+// isEmpty gets whether the specified object is considered empty or not.
+func isEmpty(object interface{}) bool {
+	// get nil case out of the way
+	if object == nil {
+		return true
+	}
+
+	objValue := reflect.ValueOf(object)
+
+	switch objValue.Kind() {
+	// collection types are empty when they have no element
+	case reflect.Chan, reflect.Map, reflect.Slice:
+		return objValue.Len() == 0
+	// pointers are empty if nil or if the value they point to is empty
+	case reflect.Ptr:
+		if objValue.IsNil() {
+			return true
+		}
+		deref := objValue.Elem().Interface()
+		return isEmpty(deref)
+	// for all other types, compare against the zero value
+	// array types are empty when they match their zero-initialized state
+	default:
+		zero := reflect.Zero(objValue.Type())
+		return reflect.DeepEqual(object, zero.Interface())
+	}
+}
+
+// requireExactlyOneNotEmpty accepts any number of values and calls t.Fatal if
+// less or more than one is empty.
+func requireExactlyOneNotEmpty(t *testing.T, v ...any) {
+	if len(v) == 0 {
+		t.Fatal("Expected some values for requireExactlyOneNotEmpty, but received none")
+	}
+
+	empty := 0
+	for _, value := range v {
+		if isEmpty(value) {
+			empty += 1
+		}
+	}
+
+	if empty != len(v)-1 {
+		t.Fatalf("Expected exactly one value to not be empty, but found %d empty values", empty)
+	}
 }
 
 // Useless key but enough to pass validation in the API